// types/bookingData.ts
export interface BookingData {
  fullName: string;
  phone: string;
  eventDate: string;
  servingTime: string;
  readyTime: string;
  address: string;
  location: string;
  area: string;
  eventType: string;
  allowFilming: boolean;
  email: string;
  comment?: string;
  downpaymentUrl?: string;
}